Supported `Resource` corresponds to each controllers.
Resource is an object constructed by a Request.
It will have one or more properties to take some resources you like
such as rows in DB, file objects and so on.
All of these properties is taken by using values in Request.

This class can be specified by attribute on Controller class.

And then, the view fixed to take Request and Context (instance of resource).
accodingry it, the Predicate, View Wrappers fixed too.